+++ docassemble_webapp/docassemble/webapp/server.py
@@ -29949,29 +29949,16 @@ def manage_api():
                 flash(word("Could not create new key"), 'error')
                 return render_template('pages/manage_api.html', **argu)
             argu['description'] = Markup(
-                    f"""<div class="alert alert-success" role="alert">
-                    <p>
-                    { word(f"Make sure to copy your <strong>{ form.name.data }</strong> API key now as you will not be able to see this again.") }
-                    </p>
-
-                    Your API key: <kbd id="apiKey">{ api_key }</kbd> <button onclick="copyToClipboard()" class="btn btn-secondary btn-sm">{ word("Copy") }</button>
-                    </div>
-                    
-                    <script>
-                        function copyToClipboard() {{
-                            const apiKeyElement = document.getElementById('apiKey');
-                            const selection = window.getSelection();
-                            const range = document.createRange();
-                            range.selectNodeContents(apiKeyElement);
-                            selection.removeAllRanges();
-                            selection.addRange(range);
-                            document.execCommand('copy');
-                            selection.removeAllRanges();
-                            alert('{ word("API Key copied to clipboard!") }');
-                        }}
-                    </script>
-                    """
-            )
+                    """<div class="card text-bg-light mb-3">
+                      <div class="card-body">
+                        <p class="card-text">
+                        """ + (word("Your new API key, known internally as <strong>%s</strong>, is:<br />%s<br />") % (form.name.data, '<br /><span class="text-success"><i class="fa-solid fa-check"></i></span> <code id="daApiKey">' + api_key + '</code><wbr /><button aria-label=' + json.dumps(word("Copy API key")) + ' onclick="daCopyToClipboard()" class="btn btn-link ps-1 pt-1" type="button"><i class="fa-regular fa-copy"></i></button>')) + """
+                        </p>
+                        <p class="card-text">
+                      """ + word("<strong>This is the only time you will be able to see your API key</strong>, so make sure to make a note of it and keep it in a secure place.") + """
+                        </p>
+                      </div>
+                    </div>""")
 
         elif action == 'edit':
             argu['title'] = word("Edit API Key")